global interface: cleanup: The server side ./version file and the .e2/version file used to hold the global interface version, describing the way data is stored on server side. If the way data is stored is changed the version must be changed, too. That must take effect for all old versions of the project, so redundantly storing the global interface version in .e2/version must not be done.
This commit cleans up the situation:
remove and deprecate .e2/version
introduce .e2/global_version that is created on project check-out and
initialized with the value stored in ./version on the server side.

Local tools check the global interface version against a list of
supported versions stored in the build time configuration. That way they
will, in future, be able to complain if a particular version is not
supported.

Global tools create projects with the newest global interface version
supported.

Signed-off-by: Gordon Hecker <gh@emlix.com>