local/e2-new-source.lua: check if file is already on the server
If we try to create a new source, using the "--files" parameter
then we now check if a file with the same name is already on the
server.

Before this patch, it was possible to overwrite an existing file
without any notifications.

e2-new-source aborts now if a file with the same name already exists.

Signed-off-by: Fabian Godehardt <fg@emlix.com>