On glibc, strerror_r does not do what you think it does...
Fixes empty error messages in various places.

glibc will use the GNU version of strerror_r by default, which does not
fill the buffer but rather uses it as a temporary storage space. It
returns a pointer to the error string, which was never used.

Since e2 is not multithreaded, replace it with the simpler strerror()
and remove pointless casts.

Signed-off-by: Tobias Ulmer <tu@emlix.com>