Try to make --log-debug documentation less confusing
This is a developer option with limited use to the general user. It
makes it possible to see at which warning or log level a message is sent
to the console.

This information is also available in debug.log, thus this option is more
or less obsolete. Too bad it's documented already...

Signed-off-by: Tobias Ulmer <tu@emlix.com>