cache: rework the caching strategy
Introduce new functions to fetch and push files in a consistent way.

To query server flags we now have:

- cache.cache_enabled()
- cache.islocal_enabled()
- cache.writeback_enabled()

All take the optional flags argument, allowing for overrides when
necessary.

The flags table is now checked for unknown flags, throwing a runtime
error. Unused flags have been removed.

Removed the function to cache files explicitly, return the file path if
in cache, and removed some dead code.
If the cache is enabled, calls to any function requesting path or a
copy of the requested file will cache automatically. This behaviour can
now be suppressed with the flags argument.

Introduced a function to check for the existence of a file on the remote
side without downloading:
transport.file_exists()

Added new function
- cache.fetch_file_path()
that returns the requested files path to either the cache, local
file system or a temporary location.

The code base was adjusted to work without any caching if desired,
although this is not a recommended mode of operation due to speed and
temporary storage requirements.

Signed-off-by: Tobias Ulmer <tu@emlix.com>