Add experimental gitrepo source plugin
Works just like a git source, but provides the entire git repository to
the build environment. This eliminates the need for wrapping a git
repository in a file source. Checkout of the configured branch of tag is
done automatically.

The configuration looks just like a "git" source, except its type must be
"gitrepo".

The plugin is currently experimental and not recommended for production
use.

Massaged heavily by tu@

Signed-off-by: Tobias Ulmer <tu@emlix.com>