e2lib: fix e2lib.wait() for signaled process
wait() returned 0 for a signaled process, expecting the caller to
check the third argument for a signal code.

Most call sites do not observe the signal argument, leading to
"successful" execution of a crashing process.

wait() will now return the exit code or signal + 128 in the first
argument. This is in line with various shells.

Also improve debug logging around wait() and callcmd().

Signed-off-by: Tobias Ulmer <tu@emlix.com>