git: optionally skip workdir checkout
Skips workdir checkout on git clone, like in older versions of
e2factory.  This may help to speed up fetch-sources of very large repos
and reduces disk space requirements in some scenarios (CI).

It's currently not documented and may go away. For one, configuring it
via the source config is a little awkward.

Signed-off-by: Tobias Ulmer <tu@emlix.com>