e2build: pty use breaks fg session signal handling, revert
Signal handling involving a pty and a foreground process inside the
chroot turns out to be very tricky. Having the ability to terminate suid
e2-su and also send eg. ^C to a vim inside doesn't appear to be
documented much. There are a few hints in sudo and related MLs.

Since it's more important to interact with applications in playground
than being able to reliably terminate e2-su, opt for a non-pty
controlled playground environment for the time being.

Signed-off-by: Tobias Ulmer <tu@emlix.com>