e2tool: file_class: add hashupdate functionality
This change provides the groundwork for files on servers other than "."
to change and have their sha1/sha256 checksums re-computed on the fly.

It will also invalidate the cached version of the file. If the server
transport has ssh support, it will first attempt to calculate the
checksum remotely, otherwise it will have to re-fetch the file every
time.

If hashupdate is used in conjunction with sha1/sha256, it will work much
like --check-remote option and complain if there is a checksum mismatch.

Signed-off-by: Tobias Ulmer <tu@emlix.com>